Feedback

+++ src/api/types.h
@@ -1,7 +1,8 @@
 #pragma once
 
 #include <memory>
-#include "../types/gamestate.h"
+#include "types/gamestate.h"
+#include "stdint.h"
 
 namespace api {
 
@@ -18,7 +19,7 @@ typedef struct CEvent {
 };
 
 typedef struct CGameSettings {
-  unsigned int seed;
+  uint64_t seed;
   const char* seat_controllers[4];
   int num_controllers = 0;
 };
